Fix major compilation errors - Part 4
- Removed duplicate class definitions from Navigation.kt (NavigationSettings, ParkingInfo, GasStationInfo, RouteEvent)
- Fixed destStr_ typo to destStr in RouteCache.kt
- Added RouteType and OsmGeoPoint imports to NavigationManager
- Fixed Double to Float conversion in AIRoutePredictor.calculateTrafficScore
- Cleaned up all duplicate data class redeclarations
- All critical type mismatches and redeclaration errors resolved

Remaining issues to fix in next commit:
- Activity binding errors (DashboardActivity, MusicActivity, NavigationActivity)
- ProfessionalNavigationActivity navigation and binding issues
- VoiceNavigationHelper steps property access
- Minor LatLng type mismatches in NavigationManager

+++ app/src/main/java/com/persianai/assistant/navigation/ai/AIRoutePredictor.kt
@@ -190,7 +190,7 @@ class AIRoutePredictor(private val context: Context) {
      */
     private fun calculateTrafficScore(route: NavigationRoute, trafficFactor: Float): Float {
         // مسیرهای طولانی‌تر در ترافیک بیشتر آسیب می‌بینند
-        return route.distance * trafficFactor
+        return route.distance.toFloat() * trafficFactor
     }
     
     /**
